app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.43
asyncpg==0.29.0
google-generativeai==0.8.3
pydantic==2.11.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.28.1
